use size_t instead of int in geometry analytic

+++ src/t8_geometry/t8_geometry_implementations/t8_geometry_analytic.hxx
@@ -46,7 +46,7 @@
 typedef void        (*t8_geom_analytic_fn) (t8_cmesh_t cmesh,
                                             t8_gloidx_t gtreeid,
                                             const double *ref_coords,
-                                            const int num_coords,
+                                            const size_t num_coords,
                                             double out_coords[3],
                                             const void *tree_data,
                                             const void *user_data);
@@ -64,7 +64,7 @@ typedef void        (*t8_geom_analytic_fn) (t8_cmesh_t cmesh,
 typedef void        (*t8_geom_analytic_jacobian_fn) (t8_cmesh_t cmesh,
                                                      t8_gloidx_t gtreeid,
                                                      const double *ref_coords,
-                                                     const int num_coords,
+                                                     const size_t num_coords,
                                                      double *jacobian,
                                                      const void *tree_data,
                                                      const void *user_data);
